[MSE] Fix erase range to prevent accidental deletion in files with changing durations https://bugs.webkit.org/show_bug.cgi?id=233528
Reviewed by NOBODY (OOPS!).

Source/WebCore:

In didReceiveSample(), eraseBeginTime was being set to
highestPresentationTimestamp minus a tolerance. The tolerance is not
needed since highestPresentationTimestamp is loaded from exact frame
timestamps, and can cause accidental frame erasure in situations where
there are frames with frames smaller than the tolerance, which is the
case for certain MP4 files.

Test: media/media-source/media-source-append-tiny-durations.html

* platform/graphics/SourceBufferPrivate.cpp:
(WebCore::SourceBufferPrivate::didReceiveSample):

LayoutTests:

* media/media-source/media-source-append-tiny-durations-expected.txt: Added.
* media/media-source/media-source-append-tiny-durations.html: Added.