[GStreamer][MSE] Preserve durations when using parsers
Parser elements like h264parse are engaged after demuxers to correct missing
metadata such as what frames are sync samples. Files with omissions like these
are common for auto-generated transmuxed content.

Unfortunately, in the case of video, parser elements generally don't take into
consideration the duration of the input buffers, instead guessing it from the
caps or leaving it unspecified. This is particularly a problem for files where
the frame durations are not constant, leading to innacurate timestamps and
unexpected gaps and overlaps.

Varying frame durations are sometimes used in MP4 to avoid having a one
frame-long gap in content using B-frames, as an alternative to fixing that with
edit lists. Note the durations encoded in MP4 exist in the decode timeline
(decode durations), therefore the outputted PTS can (and usually still are)
evenly spaced.

[Theory addendum: The time between consecutive frames in the presentation
timeline in MP4 is sometimes referred to as presentation duration, but this is
a derived concept and computing those is non-causal (requires knowledge of
future frames) and therefore not viable to compute in MSE.]

* platform/graphics/gstreamer/mse/AppendPipeline.cpp:
(WebCore::AppendPipeline::AppendPipeline):
(WebCore::ParserDurationPreservingProbes::setUp):
(WebCore::ParserDurationPreservingProbes::ParserDurationPreservingProbes):
(WebCore::ParserDurationPreservingProbes::sinkPadProbeCallback):
(WebCore::ParserDurationPreservingProbes::srcPadProbeCallback):
(WebCore::createOptionalParserForFormat):
(WebCore::AppendPipeline::Track::initializeElements):
* platform/graphics/gstreamer/mse/AppendPipeline.h: