watchOS arm64_32 build broken due to ThreadSafeRefCounted layout https://bugs.webkit.org/show_bug.cgi?id=279227 rdar://problem/135371304
Reviewed by Alex Christensen.

With assertions enabled, WTF::RefCounted has m_isOwnedByMainThread,
m_areThreadingChecksEnabled, and m_adoptionIsRequired boolean fields
which implement runtime assertion checks. On platforms with 8-byte
pointers, these are packed in to the object without affecting the
overall size of the class, but on arm64_32 they make it a byte larger.
This causes a GeneratedSerializers.mm assertion which compares the sizes
of WTF::ThreadSafeRefCounted and WTF::RefCounted to fail:

    error: static assertion failed due to requirement 'sizeof(ShouldBeSameSizeAsThreadSafeDataBufferImpl) == sizeof(WebCore::ThreadSafeDataBufferImpl)'
     31638 |     static_assert(sizeof(ShouldBeSameSizeAsThreadSafeDataBufferImpl) == sizeof(WebCore::ThreadSafeDataBufferImpl));
    GeneratedSerializers.mm:31638:70: note: expression evaluates to '28 == 24'
     31638 |     static_assert(sizeof(ShouldBeSameSizeAsThreadSafeDataBufferImpl) == sizeof(WebCore::ThreadSafeDataBufferImpl));

To fix this without teaching IPC serializers about the different
implementations of refcounting, pad ThreadSafeRefCounted with the same
bool fields to get the expected layout.

* Source/WTF/wtf/Compiler.h: Add UNUSED_MEMBER_VARIABLE (unlike
  WK_UNUSED_INSTANCE_VARIABLE, it it not disabled in some API/SPI
  headers)
* Source/WTF/wtf/RefCounted.cpp:
* Source/WTF/wtf/ThreadSafeRefCounted.h:

Canonical link: https://commits.webkit.org/283247@main