ciso646:46:4: warning: "<ciso646> is deprecated in C++17, use <version> to detect implementation-specific macros" [-W#warnings] https://bugs.webkit.org/show_bug.cgi?id=286324
Reviewed by Adrian Perez de Castro.

The purpose of this header is to access macros like __GLIBC__. It is
obsoleted and replaced by <version> but we generally don't need to
include this manually because it will be pulled in by other system
headers. It was originally needed here to support the DisallowCType
check that was added in 96006@main and removed in 288854@main, so
probably don't need it anymore.

* Source/WebCore/PAL/config.h:
* Source/WebCore/config.h:

Canonical link: https://commits.webkit.org/289666@main

(cherry picked from commit 403fba703947e7b9e7bb9569f43b5d6a45e8ce49)