[Tools][WPE] run-mvt-tests should print some status updates meanwhile is running the suites https://bugs.webkit.org/show_bug.cgi?id=288549
Reviewed by Philippe Normand.

The run-mvt-tests script is not printing any status update meanwhile
the suites are run, and sometimes some of those suites can take more
than 20 minutes (1200 seconds), so that is causing random problems
on the buildbot because buildbot is configured to kill any test step
that spends more than 1200 seconds without printing anything.

To avoid this problem, and also to improve the feedback received when
running the test suites, this patch changes to code to print each 5
seconds an update of the progress on the run of the test suite.

This is printed on the same line, with \r (carriage return), so this
status updates won't appear on the final output, only when running
the test suite.

* Tools/Scripts/run-mvt-tests:

Canonical link: https://commits.webkit.org/291107@main

(cherry picked from commit 940df310a954df4bbcbdb20c86034bfa54086eef)