[Tools][WPE] run-mvt-tests should retry unexpected failures https://bugs.webkit.org/show_bug.cgi?id=289349
Reviewed by Adrian Perez de Castro.

Some of the MVT tests are randomly flaky, and this is causing issues
on the bot runners due this random unexpected failures.

This patch modifies the runner so each unexpected failure is retried
up to N times (default 3) before marking it as such.

It also improves the conditions to check when the tests have started
or finished. Previously there was a bug detecting that all the tests
finished completely. It also improves the logging to summarize at the
end of each step the names of the tests that failed (expected or not).

* Tools/Scripts/run-mvt-tests:
(parse_args):

Canonical link: https://commits.webkit.org/292058@main

(cherry picked from commit 3e12a025ba711af6d7027624fbcc7cc2154ba9f3)