Allow EmptyOriginAccessPatterns singleton to be used on the WebProcess too.
The WebProcess should use the OriginAccessPatternsForWebProcess singleton to
handle its access patterns, and the EmptyOriginAccessPatterns is meant to be
used just by the rest of the processes. But when implementing the whitelist
API, we added a check where we call SecurityOrigin::isSameOriginDomain() to
find whether a domain access is whitelisted where we don't need to check the
WebProcesss access patterns but just the added allowed mappings. For that case,
using OriginAccessPatternsForWebProcess is computational waste, as those checks
have already been done, so we use the EmptyOriginAccessPatterns instead.