Avoid round-up of small volume capacities
For use cases where the volume capacity override option cannot
be used, if we round small capacities to the next 1GB, we might
use inadequate values for proper origin and total storage quota
control. This is more critical for smaller capacities.

Note: It is arguable if we should be rounding up at all, because
that will lead to the use of wrong quotas. Rounding down would
seem more sensible to ensure we won't attempt to use more space
than the one available. For now, adjust for the small volume
sizes