[GTK][WPE] Stuttering when playing WebP animations https://bugs.webkit.org/show_bug.cgi?id=297393
Reviewed by Carlos Garcia Campos.

When clearing the FrameBuffer cache, keep the most recent complete frame
around, because that will probably be necessary to decode the next frame.
This avoids having to decode several frames in a row in order to get the
next frame after a call to clearFrameBufferCache() is performed.

* Source/WebCore/platform/image-decoders/webp/WEBPImageDecoder.cpp:
(WebCore::WEBPImageDecoder::clearFrameBufferCache):

Canonical link: https://commits.webkit.org/298831@main