Don't use m_shader cache and create a new one for each draw https://bugs.webkit.org/show_bug.cgi?id=298220
Reviewed by Carlos Garcia Campos.

The usage of `m_shader` doesn't seem to be safe. It seems to me that a Gradient
might be used by different threads. They both can check that the shader is
not set and then try to create it. The first one creates it and then tries to
use it and the second one will then create a new one and free the first one.
I could see two threads that use `WebCore::Gradient::fill()` which ended
up in `SkShaderBase::makeContext()` crash.

This commit removes the `m_shader` cache and always creates a new shader for
each draw. This is probably more expensive, but maybe not that much.

If I remove the use of the shader cache, the crash goes away.

* Source/WebCore/platform/graphics/Gradient.h:
* Source/WebCore/platform/graphics/skia/GradientSkia.cpp:
(Gradient::shader): Modified

Canonical link: https://commits.webkit.org/299797@main